+++ src/components/buisness/mypage/UserDrawerTaps.tsx
@@ -9,7 +9,7 @@ import {
 import { useEffect, useState } from 'react';
 import { getFavoriteWebtoonList } from '@/lib/api/webtoon/favorite';
 import { WebtoonDetailDto } from '@/lib/types/webtoon/WebtoonDetailDto';
-import WebtoonList from '@/components/common/webtoon/WebtoonList';
+import WebtoonList from '@/components/common/WebtoonList/WebtoonList';
 
 const UserDrawerTaps = () => {
   const [favoriteWebtoons, setFavoriteWebtoons] = useState<WebtoonDetailDto[]>(
