+++ src/components/common/WebtoonList/WebtoonList.tsx
@@ -1,4 +1,4 @@
-import WebtoonItem from '@/components/common/Webtoon/WebtoonItem';
+import WebtoonItem from '@/components/common/WebtoonList/WebtoonItem';
 import { WebtoonDetailDto } from '@/lib/types/webtoon/WebtoonDetailDto';
 
 interface WebtoonListProps {
